move sequence module to track.
