+++ opentimelineio/schema/track.py
